Main

+++ scripts/update_rh.sh
@@ -30,11 +30,15 @@ if [ -d "/home/${1}/RotorHazard-*" ]; then
 fi
 sudo rm -r /home/"${1}"/temp.zip >/dev/null 2>&1 # in case of weird sys config or previous unsuccessful installations
 cd /home/"${1}" || exit
-wget https://codeload.github.com/RotorHazard/RotorHazard/zip/"${2}" -O temp.zip
-unzip temp.zip
-rm ~/wget* >/dev/null 2>&1
-mv /home/"${1}"/RotorHazard-* /home/"${1}"/RotorHazard
-sudo rm temp.zip
+if [ "$3" == "git" ]; then
+  git clone -c advice.detachedHead=false -b "${2}" https://github.com/RotorHazard/RotorHazard.git
+else
+  wget https://codeload.github.com/RotorHazard/RotorHazard/zip/"${2}" -O temp.zip
+  unzip temp.zip
+  rm ~/wget* >/dev/null 2>&1
+  mv /home/"${1}"/RotorHazard-* /home/"${1}"/RotorHazard
+  sudo rm temp.zip
+fi
 sudo mkdir /home/"${1}"/backup_RH_data >/dev/null 2>&1
 sudo chmod 777 -R /home/"${1}"/RotorHazard/src/server
 sudo chmod 777 -R /home/"${1}"/RotorHazard_"${upgradeDate}"
@@ -52,8 +56,6 @@ cp /home/"${1}"/RotorHazard_"${upgradeDate}"/src/server/database.db /home/"${1}"
 cp /home/"${1}"/RotorHazard_"${upgradeDate}"/src/server/database.db /home/"${1}"/backup_RH_data >/dev/null 2>&1 &
 time_warning_show
 cd /home/"${1}"/RotorHazard/src/server || echo "$red missing RotorHazard directory"
-python3 -m venv venv
-source venv/bin/activate
 pip3 install --upgrade pip
 pip3 install --upgrade --no-cache-dir -r requirements.txt
 pip3 install cffi pillow
@@ -83,7 +85,7 @@ if ! test -f "$PYTHON3_CONVERSION_FLAG_FILE"; then
 
   printf "\n"
 
-  if grep -Fq "python server.py" "/home/"${1}"/.bashrc"; then
+  if grep -Fq "python server.py" "/home/'${1}'/.bashrc"; then
     echo "old python based server-start alias found"
     sed -i 's/python server.py/python3 server.py/g' ~/.bashrc
     echo "'ss' alias changed to python3 version"
